Add files via upload
Improved transfer learning model(fine-tuned it) and added K-fold cross validation for the CNN. Will add the K-fold validation for the transfer learning model tomorrow, it is running on hipergator. I don't know if this is necessary I think it's only needed for simple models like random forest but I'm adding it anyways.